Temporary fixes to payment group
The lines on the journal got repeated. I tracked the problem to  account_payment

https://github.com/odoo/odoo/blob/80bfcde4eafefd4d08b562aba7fbfec83e23f91e/addons/account/models/account_payment.py#L605

Send one. I ended up calling _set_next_sequence on the payment movement.
The journals are not creating the right ids for their entries because the grouping they don't get posted right away  It doesn't fall on the account move . So I just hardcoded a workaround